feat: Refactor backend packages and update README
