+++ backend/src/main/java/com/eventra/controller/StatusController.java
@@ -1,4 +1,4 @@
-package com.eventra;
+package com.eventra.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
